refactor: extract header-wrapper
